Add game function and change declareWinner function to playRound
The game function play the rock, paper, scissors game 5 times and
keeps the score of the computer and the player. Finally, it reports
a winner according to the scores of five rounds.

The name of declareWinner function was changed to playRound since
the winner will be declared after playing 5 rounds of the game and
was not representing what's actually happening properly.